backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0 
psycopg2==2.9.9 
django-cors-headers==3.8.0
pytest-django==4.8.0  
celery==5.3.6 
redis==5.0.2 
drf-yasg==1.21.7 
